Don't allow whacking hidden moles.
If you aimed just right, you could click on
a mole even when it wasn't visible.
If you were fast enough, you could click on
a mole multiple times to get multiple points.

To fix both of these, only allow clicking a mole
when the hole is actually in the "show" state,
and make the mole "gone" after you click on it.